+++ babble-react/src/lookup/index.js
@@ -1,0 +1,5 @@
+import {loadTweets} from './components'
+
+export {
+    loadTweets,
+}
